+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added support for detecting separate modifier keys for terminals that support the Kitty key protocol
 - Added `TEXTUAL_DISABLE_KITTY_KEY` env var to disable Kitty key protocol support (debug aid).
 
+### Changed
+
+- Undo/redo/copy/cut/paste in TextAre will now work with cmd+ on supported terminals
+- In TextArea, ctrl+u will now delete a newline if the cursor is at the start
+
 ## [8.2.6] - 2026-04-13
 
 ### Fixed