+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased 
+## [8.2.7] - 2026-05-19 
 
 ### Added
 
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed text opacity with ANSI themes creating RGB colors.  https://github.com/Textualize/textual/pull/6546
 
-## [8.2.6] - 2026-04-13
+## [8.2.6] - 2026-05-13
 
 ### Fixed
 
@@ -3464,6 +3464,7 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[8.2.7]: https://github.com/Textualize/textual/compare/v8.2.6...v8.2.7
 [8.2.6]: https://github.com/Textualize/textual/compare/v8.2.5...v8.2.6
 [8.2.5]: https://github.com/Textualize/textual/compare/v8.2.4...v8.2.5
 [8.2.4]: https://github.com/Textualize/textual/compare/v8.2.3...v8.2.4